Move Point3D out of Day19
Move Point3D out of Day19 so it can be re-used.
